E5: fittest propaga en Double termino a termino; ronda no reabre tras stop
- Fittest (main.frm:3009): `s = score(...) + .nrg + .body*10` encadena en
  Double desde el primer operando — solo body*10 es producto Single (a
  diferencia de InvestedEnergy, que suma en Single). Corregido el orden.
- worker.js: con parada del core y StartAnotherRound colgados del mismo
  Countpop (ganador declarado), el original queda detenido en el mundo
  final — no se abre otra ronda.

Suite 165/3333 en verde en los tres modos.

+++ port/core/include/dbcore/gamemodes.hpp
@@ -104,8 +104,12 @@ inline int Fittest(Sim& sim) {
     if (b.exist && !b.Veg && b.FName != "Corpse" && !BaseHidden(sim, b)) {
       vb_long TotalOffspring = 1;
       bool Cancer = false;
+      // s = score(...) + .nrg + .body*10 — el primer operando es Double:
+      // la cadena propaga en doble termino a termino (solo body*10 es
+      // producto Single), a diferencia de InvestedEnergy (suma Single).
       double s = score0(sim, t, 1, 10, TotalOffspring, Cancer) +
-                 static_cast<double>(b.nrg + b.body * 10.0f);
+                 static_cast<double>(b.nrg) +
+                 static_cast<double>(b.body * 10.0f);
       if (s < 0) s = 0;
       s = std::pow(static_cast<double>(TotalOffspring), sPopulation) *
           std::pow(s, sEnergy);
